Pass context to type fns. Replace extended config with extra prop.
Context, previously only used for filters (and called TFilterContext)
is now also passed to type converter functions as the second argument.
This allows type conversions that depend on context (e.g. member id to
member object).

BREAKING CHANGE: Instead of allowing to specify command config type
directly, you can now specify the type of an "extra" property in the
config. This allows passing extra data with the command, which was the
original intention of allowing to specify the config type.